Addressing PR 14 feedback
