+++ nuget.config
@@ -9,7 +9,6 @@
     
     <!-- Add the SSMS repo for private requirements -->
     <add key="SQLDS - SSMS" value="http://SQLISNuget/DS-SSMS/nuget/" />
-    <add key="CrossPlat" value="W:/" />
   </packageSources>
 
 </configuration>
