+++ src/Microsoft.SqlTools.ServiceLayer/LanguageServices/AutoCompleteService.cs
